Moved Netty related classes to same package
