+++ src/main/java/com/hivemq/client/internal/mqtt/handler/auth/MqttConnectAuthHandler.java
@@ -26,8 +26,8 @@ import com.hivemq.client.internal.mqtt.message.auth.MqttEnhancedAuthBuilder;
 import com.hivemq.client.internal.mqtt.message.connect.MqttConnect;
 import com.hivemq.client.internal.mqtt.message.connect.MqttStatefulConnect;
 import com.hivemq.client.internal.mqtt.message.connect.connack.MqttConnAck;
+import com.hivemq.client.internal.netty.DefaultChannelOutboundHandler;
 import com.hivemq.client.internal.util.Checks;
-import com.hivemq.client.internal.util.netty.DefaultChannelOutboundHandler;
 import com.hivemq.client.mqtt.exceptions.ConnectionFailedException;
 import com.hivemq.client.mqtt.mqtt5.Mqtt5ClientConfig;
 import com.hivemq.client.mqtt.mqtt5.auth.Mqtt5EnhancedAuthMechanism;
