+++ src/main/java/com/hivemq/client/internal/mqtt/handler/publish/outgoing/MqttOutgoingQosHandler.java
@@ -42,11 +42,11 @@ import com.hivemq.client.internal.mqtt.message.publish.pubcomp.MqttPubComp;
 import com.hivemq.client.internal.mqtt.message.publish.pubrec.MqttPubRec;
 import com.hivemq.client.internal.mqtt.message.publish.pubrel.MqttPubRel;
 import com.hivemq.client.internal.mqtt.message.publish.pubrel.MqttPubRelBuilder;
+import com.hivemq.client.internal.netty.ContextFuture;
+import com.hivemq.client.internal.netty.DefaultContextPromise;
 import com.hivemq.client.internal.util.Ranges;
 import com.hivemq.client.internal.util.UnsignedDataTypes;
 import com.hivemq.client.internal.util.collections.IntMap;
-import com.hivemq.client.internal.util.netty.ContextFuture;
-import com.hivemq.client.internal.util.netty.DefaultContextPromise;
 import com.hivemq.client.mqtt.datatypes.MqttQos;
 import com.hivemq.client.mqtt.exceptions.ConnectionClosedException;
 import com.hivemq.client.mqtt.mqtt5.advanced.interceptor.qos1.Mqtt5OutgoingQos1Interceptor;
